Program complete, added additional test file
Program is pretty much done, but more can be added to make it even better.
Ideas for future improvement:
1: Text file with the names of all grammers and the ability to fork off
multiple process and compute them at the 'same' time.
2: better error checking. Not all edge cases were handled and feeding in a
bad file name will throw and error.
3: Better handleing of the file to inputthe CFG from. Currently the setup
is a little strange, but it works (which is what is most important right
now)
4: OOP of the LLLs. It seems silly to have implimented the same data
structure twice when they are used so simiularly. This would also be a
good chace to learn OOP in python.